Refactored TensionVector into its own file.
