+++ test.js
@@ -1,6 +1,3 @@
-import Vector from './vector'
-import is from './is'
-
 let vector1 = new Vector(1,2,3)
 let vector2 = new Vector(4,5,6)
